% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADlibrary.R
\name{MADlibrary}
\alias{MADlibrary}
\title{Installs and then loads all MADWorld packages}
\usage{
MADlibrary(
  packages = c("stable", "prototype", "world"),
  install = TRUE,
  force.install = FALSE
)
}
\arguments{
\item{packages}{Which packages to be loaded or installed. One of
"stable" (DEFAULT) for only stable packages, "prototype" for
only prototype packages, and "world" for all packages}

\item{install}{Whether to install packages that can't be loaded
(\code{TRUE} by default)}

\item{force.install}{Whether to (re-)install packages regardless
of whether they can or can't be loaded (\code{FALSE} by
default)}
}
\description{
When run with defaults, it will load (and install if necessary) all
MADworld packages. This is used to show MADworld status at package
load.
}
\examples{
\dontrun{
# What you saw when you loaded MADworld
MADlibrary("stable", install=FALSE)
# Install all MADworld packages
MADlibrary("world")
}
}
\author{
Will Pearse
}
